webapi/webapi-iap-xwalk-tests/iap/Navigator_iap_queryProductsInfo_invalidstate.html

<!DOCTYPE html>
<!--
Copyright (c) 2016 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-->

<meta charset="utf-8">
<title>WebAPI: IAP</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="He Yue" href="mailto:yue.he@intel.com">
<link rel="help" href="http://www.w3.org/TR/IAP-api/">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="options.js"></script>
<div id="log"></div>
<script>
  var at = async_test("Check if queryProductsInfo throw InvalidStateError if iap not init", {timeout: 10000});
  if (options.channel == "Google") {
    at.step(function() {
      var productsId  = ["org.xwalk.apple"];
      navigator.iap.queryProductsInfo(productsId).then(function(products) {
        assert_unreached("Here should neve be reached!");
        at.done();
      }).catch(function(e) {
        assert_equals(e.name, "InvalidStateError", "an error occurs during query raises an InvalidStateError");
        at.done();
      }); // navigator.iap.queryProdcutsInfo ends
    }); // at.step ends
  } else if (options.channel == "XiaoMi") {
    at.step(function() {
      navigator.iap.queryProductsInfo("com.demo_1").then(
        at.step_func(function() {
          assert_unreached("IAP API on Xiaomi Store does not support queryProductsInfo!");
          at.done();
        }),
        at.step_func(function(e) {
          at.done();
        })
      );
    });
  }
</script>
